9-18

+++ V1.2(recv rud angle only)/main.c
@@ -775,8 +775,8 @@ int main()
             // sendtoClient[2] = (int)(rud_send * 10);
             sendtoClient[3] = ldirrev_send + ldirrev_delta;
             sendtoClient[4] = rdirrev_send + rdirrev_delta;
-            sendtoClient[5] = ((rud_send*10) & 0xFF00) >> 8;
-             sendtoClient[6] = (rud_send*10) & 0xFF;
+            sendtoClient[5] = ((UINT32)(rud_send*10) & 0xFF00) >> 8;
+             sendtoClient[6] = (UINT32)(rud_send*10) & 0xFF;
             if(recv_joystick[0] == 0xFF)
             {
                 sendtoClient[7] = recv_joystick[3]; // x
